fixed signed integer consistency problems (i32 vs u32 vs usize) and made Ticka Field accessable.

+++ src/game_objects.rs
@@ -75,7 +75,7 @@ pub fn spawn_tower(
     let cube = Cube::new(0.1);
     let cube_mesh = meshes.add(cube.into());
 
-    let c = hex2d::Coordinate::new(x, y);
+    let c = hex2d::Coordinate::new(x as i32, y as i32);
     let (x_pixel, z_pixel) = c.to_pixel(game.hex_spacing);
 
     let material = get_color_material(materials, Color::AQUAMARINE);
